+++ gridTools/mkGridsExample3.py
@@ -28,7 +28,7 @@ grd.printMsg("At this point, we have initialized a GridUtils() object.")
 grd.printMsg("")
 
 # We can turn on extra output from the module
-grd.printMsg("We set print and logging messages to the DEBUG level.")
+grd.printMsg("Set print and logging messages to the DEBUG level.")
 logFilename = 'configs/test/nikiTest.log'
 grd.setVerboseLevel(logging.DEBUG)
 grd.setDebugLevel(0)
@@ -69,11 +69,11 @@ grd.printMsg("")
 #}, subKey='projection')
 
 # This forms a grid in memory using the specified grid parameters
-grd.printMsg("We make a grid with the grid parameters.")
+grd.printMsg("Make a grid with the grid parameters.")
 grd.makeGrid()
 
 # Save the new grid to a netCDF file
-grd.printMsg("We attempt to save the grid to a netCDF file.")
+grd.printMsg("Attempt to save the grid to a netCDF file.")
 grd.saveGrid(filename="configs/test/gridTest.nc")
 
 # This prints out all the current grid parameters
@@ -93,7 +93,7 @@ grd.printMsg("%s" % (grd.grid))
 grd.printMsg("")
 
 # Define plot parameters so we can see what the grid looks like
-grd.printMsg("We now setup plotting parameters for showing the grid on a map:")
+grd.printMsg("Setup plotting parameters for showing the grid on a map:")
 grd.setPlotParameters(
     {
         'figsize': (8,8),
@@ -130,7 +130,7 @@ grd.setPlotParameters(
 #   titles, axis, etc prior to the final plotting of the figure.
 #   Some items may be configured via the figure object.
 grd.printMsg('''
-We place a call to actually plot the grid using plotGrid().  This function returns
+Place a call to actually plot the grid using plotGrid().  This function returns
 a figure and axes object that can be further modified before displaying or saving
 the plot.
 
@@ -139,7 +139,7 @@ the plot.
 
 # You can save the figure using the savefig() method on the
 # figure object.  Many formats are possible.
-grd.printMsg("We save the figure in two different formats: jpg and pdf.")
+grd.printMsg("Save the figure in two different formats: jpg and pdf.")
 figure.savefig('configs/test/gridTest.jpg', dpi=None, facecolor='w', edgecolor='w',
         orientation='portrait', transparent=False, bbox_inches=None, pad_inches=0.1)